refactor: remove ecstore root rpc facade
