+++ crates/ecstore/src/cluster/rpc/client.rs
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use crate::cluster::rpc::{TONIC_RPC_PREFIX, gen_signature_headers};
 use crate::disk::error::{DiskError, Error as DiskErrorType};
-use crate::rpc::{TONIC_RPC_PREFIX, gen_signature_headers};
 use crate::runtime::sources as runtime_sources;
 use http::Method;
 use rustfs_protos::{create_new_channel, proto_gen::node_service::node_service_client::NodeServiceClient};
