+++ crates/ecstore/src/rebalance/rebalance_unit_tests.rs
@@ -2413,7 +2413,7 @@ async fn test_init_and_start_rebalance_rejects_second_start_after_gate() {
         id: uuid::Uuid::new_v4(),
         disk_map: std::collections::HashMap::new(),
         pools: Vec::new(),
-        peer_sys: crate::rpc::S3PeerSys::new(&endpoint_pools),
+        peer_sys: crate::cluster::rpc::S3PeerSys::new(&endpoint_pools),
         pool_meta: tokio::sync::RwLock::new(crate::core::pools::PoolMeta::default()),
         rebalance_meta: tokio::sync::RwLock::new(Some(active_meta)),
         decommission_cancelers: tokio::sync::RwLock::new(Vec::new()),
