--- crates/ecstore/src/rpc/mod.rs
@@ -1,24 +1,0 @@
-// Copyright 2024 RustFS Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#[cfg(test)]
-pub(crate) use crate::cluster::rpc::TcpHttpInternodeDataTransport;
-pub(crate) use crate::cluster::rpc::heal_bucket_local_on_disks;
-pub use crate::cluster::rpc::{
-    LocalPeerS3Client, PEER_RESTSIGNAL, PEER_RESTSUB_SYS, PeerRestClient, PeerS3Client, RemoteClient, RemoteDisk, S3PeerSys,
-    SERVICE_SIGNAL_REFRESH_CONFIG, SERVICE_SIGNAL_RELOAD_DYNAMIC, TONIC_RPC_PREFIX, TonicInterceptor, build_auth_headers,
-    build_internode_data_transport_from_env, gen_signature_headers, gen_tonic_signature_interceptor,
-    node_service_time_out_client, node_service_time_out_client_no_auth, verify_rpc_signature,
-};
-pub(crate) use crate::cluster::rpc::{client, context_propagation, internode_data_transport, runtime_sources};
